算法复习 多思路解决实际问题 不局限于单独的算法

+++ src/main/java/com/eron/algorithms/smartquestion/CarGroupProblems.java
@@ -83,7 +83,7 @@ public class CarGroupProblems {
 
         @Override
         public int compareTo(Car o) {  // 按照位置 降序 
-            return (int) (o.pos - this.pos);
+            return o.pos - this.pos;
         }
 
         @Override
